refactor: remane input.js to config_from_Args
