+++ index.js
@@ -1,2 +1,2 @@
-const config = require("./input");
+const config = require("./config_from_Args");
